Chore: Update pre-commit dependencies to latest
ERROR! the role 'shellcheck-install' was not found in <>
The error appears to be in '': line 5, column 7, but may
be elsewhere in the file depending on the exact syntax
problem.

Update role name to match the role path on GHA and
Jenkins jobs.

Add GHA workflow for verify jobs.
Update the tox and pre-commit dependencies.

Ref: https://github.com/ansible/molecule/issues/1042
Change-Id: I673570b57436c3474ee625524d2ae108664a9ee1
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>